feat(rust): deepen SentinelOne runtime kernel
